refactore: refatoracao denuncia

+++ Back/src/denuncias/denuncias.service.ts
@@ -8,55 +8,70 @@ export class DenunciasService{
 
     constructor (private prisma: PrismaService){}
 
-    async criarDenuncia (data: DenunciaDto){
+    async criarDenuncia (idUsuario: number, data: DenunciaDto){
         const criacaoDenuncia = await this.prisma.denuncia.create({
             data: {
-                idUsuario: data.idUsuario,
+                idUsuario: idUsuario,
                 descricao: data.descricao,
                 idCategoria: data.idCategoria,
-                mediasrc: data.mediasrc,
+                mediaSrc: data.mediaSrc,
                 anonimato: data.anonimato,
             }});
         return criacaoDenuncia;
     }
 
-    async editarDenuncia (id: number,data: edicaoDenunciaDto){
+    async editarDenuncia (id: number, idUsuario: number, data: edicaoDenunciaDto){
         const existeDenuncia = await this.prisma.denuncia.findUnique({
             where: { id },
         })
         if (!existeDenuncia) {
             throw new Error('Denúncia não encontrada!');
         }
+        
+        if (existeDenuncia.idUsuario !== idUsuario) {
+            throw new Error('Usuário não autorizado a editar esta denúncia!');
+        }
+
         return await this.prisma.denuncia.update({
             where: { id },
             data: {
                 descricao: data.descricao,
                 idCategoria: data.idCategoria,
-                mediasrc: data.mediasrc,
+                mediaSrc: data.mediaSrc,
                 anonimato: data.anonimato,
             },
         });
     }
 
-    async deletarDenuncia (id: number){
+    async deletarDenuncia (id: number, idUsuario: number){
         const existeDenuncia = await this.prisma.denuncia.findUnique({
             where: { id },
         })
         if (!existeDenuncia) {
             throw new Error('Denúncia não encontrada!');
         }
+        
+        if (existeDenuncia.idUsuario !== idUsuario) {
+            throw new Error('Usuário não autorizado a deletar esta denúncia!');
+        }
+        
         return await this.prisma.denuncia.delete({
             where: { id },
         });
     }
 
-    async desativarDenuncia (id: number){
+    async desativarDenuncia (id: number, idUsuario: number){
         const existeDenuncia = await this.prisma.denuncia.findUnique({
             where: { id },
         })
         if (!existeDenuncia) {
             throw new Error('Denúncia não encontrada!');
         }
+
+        if (existeDenuncia.idUsuario !== idUsuario) {
+            throw new Error('Usuário não autorizado a deletar esta denúncia!');
+        }
+        
         return await this.prisma.denuncia.update({
             where: { id },
             data: {
